Recovery and included links
A new version of Github Desktop required a reestablishment of the local repositories.

As a part of the recovery, I realised that previously the URLs of local and WWW links had not been embedded. This commit fixes that

A new procedure for exporting from Ulysses into a format appropriate for preserving in Github is outlined in Evernote